moving the helper / defs into a separate file
